+++ test/fcts.jl
@@ -4,9 +4,22 @@
 
         var_states = Dict{Int,Vector{Int}}()
         var_changes = Dict{Int,Vector{Tuple{Symbol,Int,Int,Int}}}()
+        activity = Dict{Int,Float64}()
         children = CS.TreeLogNode{Int}[]
-        l1 =
-            CS.TreeLogNode(0, :Open, true, 0, 0, 0, 0, 0, var_states, var_changes, children)
+        l1 = CS.TreeLogNode(
+            0,
+            :Open,
+            true,
+            0,
+            0,
+            0,
+            0,
+            0,
+            var_states,
+            var_changes,
+            activity,
+            children,
+        )
         l2 = CS.TreeLogNode(
             0,
             :Closed,
@@ -18,6 +31,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children,
         )
         @test !CS.same_logs(l1, l2)
@@ -35,6 +49,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         tln2 = CS.TreeLogNode(
@@ -48,6 +63,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         push!(children, tln1)
@@ -63,6 +79,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children[[1, 2]],
         )
         l2 = CS.TreeLogNode(
@@ -76,6 +93,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children[[2, 1]],
         )
         @test !CS.same_logs(l1, l2)
@@ -93,6 +111,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Float64}[],
         )
         tln2 = CS.TreeLogNode(
@@ -106,6 +125,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Float64}[],
         )
         push!(children, tln1)
@@ -121,6 +141,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children[[1, 2]],
         )
         l2 = CS.TreeLogNode(
@@ -134,6 +155,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children[[1]],
         )
         @test !CS.same_logs(l1, l2)
@@ -151,10 +173,23 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
-        tln11 =
-            CS.TreeLogNode(0, :Open, true, 0, 1, 0, 0, 0, var_states, var_changes, [tln13])
+        tln11 = CS.TreeLogNode(
+            0,
+            :Open,
+            true,
+            0,
+            1,
+            0,
+            0,
+            0,
+            var_states,
+            var_changes,
+            activity,
+            [tln13],
+        )
         tln12 = CS.TreeLogNode(
             0,
             :Open,
@@ -166,6 +201,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         push!(children1, tln11)
@@ -182,6 +218,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         tln21 = CS.TreeLogNode(
@@ -195,10 +232,23 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
-        tln22 =
-            CS.TreeLogNode(0, :Open, true, 0, 2, 0, 0, 0, var_states, var_changes, [tln23])
+        tln22 = CS.TreeLogNode(
+            0,
+            :Open,
+            true,
+            0,
+            2,
+            0,
+            0,
+            0,
+            var_states,
+            var_changes,
+            activity,
+            [tln23],
+        )
         push!(children2, tln21)
         push!(children2, tln22)
         l1 = CS.TreeLogNode(
@@ -212,6 +262,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children1,
         )
         l2 = CS.TreeLogNode(
@@ -225,6 +276,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children2,
         )
         @test !CS.same_logs(l1, l2)
@@ -242,6 +294,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         tln14 = CS.TreeLogNode(
@@ -255,6 +308,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         tln11 = CS.TreeLogNode(
@@ -268,6 +322,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             [tln13, tln14],
         )
         tln12 = CS.TreeLogNode(
@@ -281,6 +336,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         push!(children1, tln11)
@@ -297,6 +353,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         tln24 = CS.TreeLogNode(
@@ -310,6 +367,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         tln21 = CS.TreeLogNode(
@@ -323,6 +381,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             [tln24, tln23],
         )
         tln22 = CS.TreeLogNode(
@@ -336,6 +395,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             CS.TreeLogNode{Int}[],
         )
         push!(children2, tln21)
@@ -351,6 +411,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children1,
         )
         l2 = CS.TreeLogNode(
@@ -364,6 +425,7 @@
             0,
             var_states,
             var_changes,
+            activity,
             children2,
         )
         @test !CS.same_logs(l1, l2)