+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # ConstrainSolver.jl - Changelog
 
-## Unreleased 
+## v0.6.0 (15th of December 2020) 
 - **Dropped support for Julia v1.0 and v1.1**
+- Implementation of Activity Based Search
+  - `"branch_strategy" => :ABS`
 - Throw error if there is a dimension mismatch in a `TableSet`
 - Bugfix in `simplify!` check that one inner index isn't used more than once
 
@@ -75,7 +77,7 @@
   - Call `call_finished_pruning!(com)` after second `prune!` before backtracking
 
 ## v0.1.4 (6th of May 2020)
-- Added `is_solved_constraint` functions to check whether problem gets actually solved
+- Added `is_constraint_solved` functions to check whether problem gets actually solved
 - **Bugfixes:**
   - Correct incremental update in table constraint
   - Fixed `restore_pruning_constraint!` in table